+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 Release History
 ===============
+# 0.9.0
+* Emoji support!!
+* Minimum Ruby version bumped to 2.5.0 for Emoji support
+* Extend or exclude available characters using config arrays
+* Allow arrays in substitution keys
 
 # 0.8.0
 * Initial release